+++ packages/ramose/test/effect.test.ts
@@ -5,14 +5,20 @@
 
 import { describe, expect, test } from "bun:test";
 import { pipe as effectPipe } from "effect/Function";
+import * as EffectFunction from "effect/Function";
 import * as EffectRedacted from "effect/Redacted";
 import * as Ramose from "ramose/db";
-import { pipe, Redacted, Schema } from "ramose/effect";
+import { Function, pipe, Redacted, Schema } from "ramose/effect";
 
 describe("ramose/effect", () => {
   test("re-exports pipe as effect/Function.pipe", () => {
     expect(pipe).toBe(effectPipe);
     expect(pipe(1, (n) => n + 1, (n) => n * 2)).toBe(4);
+  });
+
+  test("re-exports Function as the effect/Function module instance", () => {
+    expect(Function).toBe(EffectFunction);
+    expect(Function.pipe).toBe(effectPipe);
   });
 
   test("re-exports Redacted as effect/Redacted", () => {